Extract NetworkAdd component
